Move OrInListSimplifier to its own file
