+++ datafusion/optimizer/src/simplify_expressions/mod.rs
@@ -17,6 +17,7 @@
 
 pub mod context;
 pub mod expr_simplifier;
+mod or_in_list_simplifier;
 mod regex;
 pub mod simplify_exprs;
 mod utils;
